refactor: header 컴포넌트 디자인 파일 분리 및 파일 적용
