+++ src/components/@common/Layout.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import styled from "styled-components";
-import Header from "./Header";
+import Header from "./header";
 import { useLocation } from "react-router-dom";
 
 interface Props {
