+++ src/components/@common/header/Header.styles.ts
@@ -1,0 +1,29 @@
+import styled from "styled-components";
+
+export const Styled = {
+  HeaderContainer: styled.header`
+    display: flex;
+    column-gap: 18px;
+    align-items: center;
+  `,
+
+  BackButton: styled.button`
+    background: none;
+    border: none;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 14px;
+    height: 14px;
+  `,
+
+  Heading: styled.h2`
+    font-size: 16px;
+
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 19px;
+
+    color: #383838;
+  `,
+};
